requirements.txt
langchain-core
python-dotenv
langgraph
duckduckgo-search
langchain-openai
langchain-community
google-search-results